Moved Mailing step to ConfigController
