+++ protected/humhub/modules/installer/controllers/ConfigController.php
@@ -10,6 +10,7 @@ namespace humhub\modules\installer\controllers;
 
 use humhub\components\access\ControllerAccess;
 use humhub\components\Controller;
+use humhub\modules\installer\forms\MailingForm;
 use humhub\modules\marketplace\Module;
 use humhub\modules\queue\driver\Sync;
 use humhub\modules\space\models\Space;
@@ -17,6 +18,7 @@ use humhub\modules\user\models\Group;
 use humhub\modules\user\models\Password;
 use humhub\modules\user\models\User;
 use Yii;
+use yii\base\BaseObject;
 use yii\base\InvalidConfigException;
 use yii\web\HttpException;
 
@@ -225,6 +227,44 @@ class ConfigController extends Controller
         } else {
             return $this->render('modules', ['modules' => $modules]);
         }
+    }
+
+    /**
+     * Configure SMTP
+     * Checking given SMTP configuration, writing them into a config file.
+     *
+     */
+    public function actionMailing()
+    {
+        $model = new MailingForm();
+
+        if ($model->load(Yii::$app->request->post()) && $model->validate())
+        {
+            if($model->sendTest)
+            {
+                try {
+                    $mail = Yii::$app->mailer->compose(['html' => '@humhub/views/mail/TextOnly'], [
+                        'message' => Yii::t('InstallerModule.base', 'Test message')
+                    ]);
+                    $mail->setTo($model->testEmailAddress);
+                    $mail->setSubject(Yii::t('InstallerModule.base', 'Test message'));
+
+                    if (!$mail->send()) {
+                        $errorMessage = Yii::t('InstallerModule.base', 'Could not send test email.');
+                    } else {
+                        $successMessage = Yii::t('InstallerModule.base', 'Email has been successfully sent to {testEmailAddress}. Check your inbox!', ['testEmailAddress' => $model->testEmailAddress]);
+                    }
+                } catch (\Exception $e) {
+                    $errorMessage = Yii::t('InstallerModule.base', 'Could not send test email.') . ' ' . $e->getMessage();
+                }
+
+            } else if ($model->save()) {
+                return $this->redirect(Yii::$app->getModule('installer')->getNextConfigStepUrl());
+            }
+        }
+
+        // Render Template
+        return $this->render('mailing', ['model' => $model, 'errorMessage' => $errorMessage ?? '', 'successMessage' => $successMessage ?? '']);
     }
 
     /**
